目的地の設定 [Feature] 主人公の自動移動システム Fixes #8
