--- life_game_app/lib/game.dart
@@ -1,51 +1,0 @@
-import 'package:flame/components.dart';
-import 'package:flame/game.dart';
-import 'package:flutter/material.dart';
-
-class LifeGame extends FlameGame with HasCollisionDetection {
-  static const double gameWidth = 800;
-  static const double gameHeight = 600;
-
-  @override
-  Future<void> onLoad() async {
-    await super.onLoad();
-
-    camera.viewfinder.visibleGameSize = Vector2(gameWidth, gameHeight);
-
-    add(MovingSquare());
-  }
-
-  @override
-  Color backgroundColor() => const Color(0xFF2C2C2C);
-}
-
-class MovingSquare extends RectangleComponent with HasGameReference<LifeGame> {
-  static const double speed = 100;
-  late Vector2 velocity;
-
-  @override
-  Future<void> onLoad() async {
-    super.onLoad();
-
-    size = Vector2(50, 50);
-    position = Vector2(100, 100);
-    paint.color = Colors.blue;
-
-    velocity = Vector2(speed, speed);
-  }
-
-  @override
-  void update(double dt) {
-    super.update(dt);
-
-    position += velocity * dt;
-
-    if (position.x <= 0 || position.x >= game.size.x - size.x) {
-      velocity.x *= -1;
-    }
-
-    if (position.y <= 0 || position.y >= game.size.y - size.y) {
-      velocity.y *= -1;
-    }
-  }
-}
